+++ clients/socks5/src/commands/init.rs
@@ -11,7 +11,7 @@ use crate::{
 };
 use clap::Args;
 use nym_bin_common::output_format::OutputFormat;
-use nym_client_core::init::client_init::{
+use nym_client_core::cli_helpers::client_init::{
     initialise_client, CommonClientInitArgs, InitResultsWithConfig, InitialisableClient,
 };
 use nym_sphinx::addressing::clients::Recipient;
