+++ clients/socks5/src/config/mod.rs
@@ -3,6 +3,8 @@
 
 use crate::config::template::CONFIG_TEMPLATE;
 use nym_bin_common::logging::LoggingSettings;
+use nym_client_core::cli_helpers::client_init::ClientConfig;
+use nym_client_core::config::disk_persistence::CommonClientPaths;
 use nym_config::{
     must_get_home, read_config_from_toml_file, save_formatted_config_to_file, NymConfigTemplate,
     DEFAULT_CONFIG_DIR, DEFAULT_CONFIG_FILENAME, DEFAULT_DATA_DIR, NYM_DIR,
@@ -15,9 +17,7 @@ use std::path::{Path, PathBuf};
 use std::str::FromStr;
 
 pub use crate::config::persistence::SocksClientPaths;
-use nym_client_core::config::disk_persistence::CommonClientPaths;
 pub use nym_client_core::config::Config as BaseClientConfig;
-use nym_client_core::init::client_init::ClientConfig;
 pub use nym_socks5_client_core::config::Config as CoreConfig;
 
 pub mod old_config_v1_1_13;
