+++ common/client-core/src/cli_helpers/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod client_init;
+pub mod client_run;
