+++ common/client-core/src/init/mod.rs
@@ -24,8 +24,6 @@ use serde::de::DeserializeOwned;
 use serde::Serialize;
 use std::sync::Arc;
 
-pub mod client_init;
-pub mod client_run;
 pub mod helpers;
 pub mod types;
