+++ common/client-core/src/lib.rs
@@ -1,5 +1,6 @@
 use std::future::Future;
 
+pub mod cli_helpers;
 pub mod client;
 pub mod config;
 pub mod error;
